chore: rename getBaseButtonStyles to getStyledBaseButtonStyles & fix children prop type
